<-client-> comment delete

+++ server/controllers/topicCommentsController.js
@@ -4,21 +4,87 @@ const path = require("path");
 const fs = require("fs");
 const {removeFile} = require("../utils/consts");
 
-const getCommentReplies = (commentsIds) => {
+const convertCommentItemFroTopic = async (commentsItem, currUser) => {
+    let convertedItem = {}
+
+    if (commentsItem.hasOwnProperty('created_by_user_id')) {
+        const user = await User.findOne({where: {id: commentsItem.created_by_user_id}})
+        convertedItem.created_by_user_name = user.name
+
+        if(currUser){
+            if(currUser.id === user.id){
+                convertedItem.editable = true
+            }
+        }
+
+    }
+
+    if (commentsItem.hasOwnProperty('text')) {
+        convertedItem.text = commentsItem.text
+    }
+
+    if (commentsItem.hasOwnProperty('topic_comment_id')) {
+        convertedItem.topic_comment_id = commentsItem.topic_comment_id
+    }
+
+    if (commentsItem.hasOwnProperty('createdAt')) {
+        convertedItem.createdAt = commentsItem.createdAt
+    }
+
+    if (commentsItem.hasOwnProperty('replies')) {
+        if (Array.isArray(commentsItem.replies)) {
+            if (commentsItem.replies.length > 0) {
+                convertedItem.replies = commentsItem.replies
+            }
+        }
+    }
+
+    return convertedItem
+}
+const getCommentReplies = async (commentsIds, topic_id, currUser, sort_code) => {
     let replies = []
     const replyIds = JSON.parse(commentsIds)
-    replyIds.map(item => {
-        const replyComment = TopicComments.findOne({
-            where: {
-                id: item
-            }
-        })
-        if(replyComment){
-            replyComment.replies = getCommentReplies(replyComment.reply_ids)
 
-            replies.push(replyComment)
+    for (let i = 0; i < replyIds.length; i++) {
+        const item = replyIds[i]
+
+        const replyComment = await TopicComments.findOne({
+            where: {
+                topic_comment_id: item,
+                topic_id
+            },
+            // order: [
+            //     sortOrder
+            //     // ['id', 'ASC'],
+            //     // ['name', 'DESC'],
+            // ]
+
+        })
+
+        if (replyComment) {
+            let newItem = JSON.parse(JSON.stringify(replyComment))
+            newItem.replies = await getCommentReplies(newItem.reply_ids, topic_id, currUser, sort_code)
+
+            replies.push(await convertCommentItemFroTopic(newItem, currUser))
         }
-    })
+    }
+
+    switch (sort_code) {
+        case 'date':
+            replies.sort(function(a, b){return a.createdAt - b.createdAt});
+            break
+        case 'redate':
+            replies.sort(function(a, b){return b.createdAt - a.createdAt});
+            break
+        case 'id':
+            replies.sort(function(a, b){return a.topic_comment_id - b.topic_comment_id});
+            break
+        case 'reid':
+            replies.sort(function(a, b){return b.topic_comment_id - a.topic_comment_id});
+            break
+    }
+
+
 
     return replies
 }
@@ -30,39 +96,63 @@ class TopicCommentsController {
             const {
                 text,
                 topic_id,
-                topic_comment_id,
-                created_by_user_id,
+                // topic_comment_id = -1,
+                // created_by_user_id,
                 on_topic_comment_reply_id = -1,
                 is_reply = false,
             } = req.body
 
-            // reply_ids: {type: DataTypes.STRING},
-            // on_topic_comment_reply_id: {type: DataTypes.INTEGER},
-            // is_reply: {type: DataTypes.BOOLEAN, defaultValue: false},
-
+            const currUser = req.user
             // file_name: {type: DataTypes.STRING},
 
-            if (text && topic_id && topic_comment_id && created_by_user_id) {
+            if (text && topic_id && currUser) {
+
+                const topicComments = await TopicComments.findAndCountAll({
+                    where: {
+                        topic_id
+                    },
+                })
+
                 const comment = await TopicComments.create({
                     text,
                     topic_id,
-                    topic_comment_id,
-                    created_by_user_id,
+                    topic_comment_id: topicComments.count,
+                    created_by_user_id: currUser.id,
                     created_date: Date.now(),
+                    on_topic_comment_reply_id,
+                    is_reply,
+                    reply_ids: '[]',
                 })
 
-                if (is_reply && on_topic_comment_reply_id > -1) {
-                    const candidate = await TopicComments.findOne({where: {id: on_topic_comment_reply_id}})
-                    if (candidate) {
-                        let replyIds = JSON.parse(candidate.reply_ids)
-                        replyIds.push(comment.id)
 
-                        candidate.reply_ids = JSON.stringify(replyIds)
-                        candidate.save()
+                if (is_reply && on_topic_comment_reply_id > -1) {
+                    const candidate = await TopicComments.findOne({
+                        where: {
+                            topic_comment_id: on_topic_comment_reply_id,
+                            topic_id: topic_id,
+                        }
+                    })
+                    if (candidate) {
+                        let replyIds = JSON.parse(candidate.reply_ids) || []
+                        replyIds.push(comment.topic_comment_id)
+
+                        // candidate.reply_ids = JSON.stringify(replyIds)
+
+                        await TopicComments.update({
+                            reply_ids: JSON.stringify(replyIds),
+                        }, {
+                            where: {
+                                id: candidate.id,
+                                // topic_comment_id: on_topic_comment_reply_id,
+                                // topic_id: topic_id,
+                            }
+                        })
+
+                        // await candidate.save()
                     }
                 }
 
-                return res.json({status: 'ok', id: comment.id})
+                return res.json({status: 'ok', id: comment.topic_comment_id})
             } else {
                 return next(ApiError.forbidden("Необходимо указать все данные комментария..."))
             }
@@ -74,24 +164,33 @@ class TopicCommentsController {
     async editComment(req, res, next) {
         try {
             const {
-                id,
                 text,
                 topic_id,
                 topic_comment_id,
-                created_by_user_id,
             } = req.body
+
+            const currUser = req.user
 
             // file_name: {type: DataTypes.STRING},
 
-            if (id && text && topic_id && topic_comment_id && created_by_user_id) {
+            if (text && topic_id && currUser) {
 
-                const candidate = await TopicComments.findOne({where: {id}})
+                const candidate = await TopicComments.findOne({
+                    where: {
+                        topic_id,
+                        topic_comment_id,
+                    }
+                })
                 if (candidate) {
-                    candidate.text = text
+                    if(candidate.created_by_user_id === currUser.id || currUser.isAdmin) {
+                        candidate.text = text
 
-                    candidate.save()
+                        candidate.save()
 
-                    return res.json({status: 'ok', id: comment.id})
+                        return res.json({status: 'ok'})
+                    }else{
+                        return next(ApiError.forbidden("Нет прав для редактирования"))
+                    }
                 }
 
                 return res.json({status: 'error', id})
@@ -104,11 +203,12 @@ class TopicCommentsController {
     }
 
     async getAll(req, res) {
-        const {sort_code, topic_id} = req.query
+        const {sort_code='reid', topic_id} = req.query
+        const currUser = req.user
 
-        if(topic_id) {
+        if (topic_id) {
 
-            let sortOrder = ['id', 'ASC']
+            let sortOrder = ['id', 'DESC']
             switch (sort_code) {
                 case 'user':
                     sortOrder = ['created_by_user_id', 'ASC']
@@ -145,32 +245,17 @@ class TopicCommentsController {
             )
             let newRows = []
 
-            commentsList.rows.map(item => {
+            for (let i = 0; i < commentsList.count; i++) {
+                const item = commentsList.rows[i]
+
                 let newItem = JSON.parse(JSON.stringify(item))
 
-                const user = User.findOne({where: {id: newItem.created_by_user_id}})
-
-                newItem.created_by_user_name = user.name
-
                 if (newItem.reply_ids) {
-                    newItem.replies = getCommentReplies(newItem.reply_ids)
-
-                    // const replyIds = JSON.parse(newItem.reply_ids)
-                    // replyIds.map(item => {
-                    //     const replyComment = TopicComments.findOne({
-                    //         where: {
-                    //             id: item
-                    //         }
-                    //     })
-                    //     if(replyComment){
-                    //         newItem.replies.push(replyComment)
-                    //     }
-                    // })
+                    newItem.replies = await getCommentReplies(newItem.reply_ids || '[]', newItem.topic_id, currUser, sort_code)
                 }
 
-                newRows.push(newItem)
-
-            })
+                newRows.push(await convertCommentItemFroTopic(newItem, currUser))
+            }
 
             return res.json({
                 count: newRows.length,
@@ -185,6 +270,8 @@ class TopicCommentsController {
 
     async deleteComment(req, res, next) {
         const {id} = req.query
+        const currUser = req.user
+
         try {
             if (!id) {
                 return next(ApiError.badRequest("Ошибка параметра"))
@@ -194,37 +281,47 @@ class TopicCommentsController {
 
                 if (candidate) {
 
-                    try {
-                        let imgFileName = candidate.file_name.split('\\')[1]
-                        const imgFilePath = path.resolve(__dirname, '..', "static", imgFileName)
-                        fs.unlinkSync(imgFilePath)
-                    } catch (e) {
-                    }
+                    if(candidate.created_by_user_id === currUser.id || currUser.isAdmin) {
 
-                    const result = removeFile(candidate.file_name)
-                    if (result.hasOwnProperty('status')) {
-                        if (result.status === 'ok') {
-
-                            await Files.destroy({where: {table_name: 'TopicComments', file_name: candidate.file_name}})
-
-                            if (candidate.is_reply && candidate.on_topic_comment_reply_id > -1) {
-                                const replyCandidate = await TopicComments.findOne({where: {id: candidate.on_topic_comment_reply_id}})
-                                if (replyCandidate) {
-                                    let replyIds = JSON.parse(replyCandidate.reply_ids)
-                                    replyIds.push(candidate.id)
-
-                                    const filtered = replyIds.filter(function (value) {
-                                        return ("" + value.id) === ("" + candidate.id)
-                                    })
-
-                                    replyCandidate.reply_ids = JSON.stringify(filtered)
-                                    replyCandidate.save()
-                                }
-                            }
-
-                            const count = await TopicComments.destroy({where: {id: id}})
-                            return res.json({status: "ok", message: `Удалено записей: ${count}`})
+                        try {
+                            let imgFileName = candidate.file_name.split('\\')[1]
+                            const imgFilePath = path.resolve(__dirname, '..', "static", imgFileName)
+                            fs.unlinkSync(imgFilePath)
+                        } catch (e) {
                         }
+
+                        const result = removeFile(candidate.file_name)
+                        if (result.hasOwnProperty('status')) {
+                            if (result.status === 'ok') {
+
+                                await Files.destroy({
+                                    where: {
+                                        table_name: 'TopicComments',
+                                        file_name: candidate.file_name
+                                    }
+                                })
+
+                                if (candidate.is_reply && candidate.on_topic_comment_reply_id > -1) {
+                                    const replyCandidate = await TopicComments.findOne({where: {id: candidate.on_topic_comment_reply_id}})
+                                    if (replyCandidate) {
+                                        let replyIds = JSON.parse(replyCandidate.reply_ids)
+                                        replyIds.push(candidate.id)
+
+                                        const filtered = replyIds.filter(function (value) {
+                                            return ("" + value.id) === ("" + candidate.id)
+                                        })
+
+                                        replyCandidate.reply_ids = JSON.stringify(filtered)
+                                        replyCandidate.save()
+                                    }
+                                }
+
+                                const count = await TopicComments.destroy({where: {id: id}})
+                                return res.json({status: "ok", message: `Удалено записей: ${count}`})
+                            }
+                        }
+                    }else{
+                        return next(ApiError.forbidden("Нет прав для удаления"))
                     }
                 }
             }
